Remove workflow changes from generated-source observability PR
